10.phidata_multiagent/requirements.txt
yfinance
phidata 
duckduckgo-search
fastapi[standard]
sqlalchemy
uvicorn
packaging
google-generativeai